Fix tray icon handling
