+++ app/build.rs
@@ -1,9 +1,9 @@
-extern crate winres;
+extern crate windres;
 
 fn main() {
-    if cfg!(target_os = "windows") {
-        let mut res = winres::WindowsResource::new();
-        res.set_icon_with_id("./res/trayIcon.ico", "trayIcon");
-        res.compile().unwrap();
+    #[cfg(target_os = "windows")]
+    {
+        use windres::Build;
+        Build::new().compile("./res/resources.rc").unwrap();
     }
 }
