+++ app/res/resources.rc
@@ -1,0 +1,1 @@
+trayIcon ICON "trayIcon.ico"
